feat: add HomeComposer component for discussion creation with tag selection and preview functionality
feat: implement UserProfileBuilders for user profile navigation and rendering helpers
