+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "ramon/avocado",
-    "version": "2.1.22-beta",
+    "version": "2.1.23-beta",
     "description": "A modern, feature-rich Flarum theme with advanced search, real-time messaging, and comprehensive multi-language support. Built on the Avocado foundation with significant enhancements.",
     "keywords": [
         "flarum",
