+++ js/src/forum/components/UserProfilePage.tsx
@@ -1,13 +1,8 @@
 // @ts-nocheck
 import app from 'flarum/forum/app';
 import UserPage from 'flarum/forum/components/UserPage';
-import Avatar from 'flarum/common/components/Avatar';
-import AvatarEditor from 'flarum/forum/components/AvatarEditor';
-import Dropdown from 'flarum/common/components/Dropdown';
 import Tooltip from 'flarum/common/components/Tooltip';
 import UserControls from 'flarum/forum/utils/UserControls';
-import listItems from 'flarum/common/helpers/listItems';
-import SelectDropdown from 'flarum/common/components/SelectDropdown';
 import {
   trans,
   displayName,
@@ -19,6 +14,7 @@ import {
 } from '../utils';
 import ThreadCard from './shared/ThreadCard';
 import PostCard from './shared/PostCard';
+import { buildHero, buildSidebar, buildUserPhoneNav } from './UserProfileBuilders';
 
 const PAGE_SIZE = 20;
 
@@ -30,232 +26,6 @@ const findBySlug = (slug: string): any => {
     (u: any) => (u.slug?.() || '').toLowerCase() === l || (u.username?.() || '').toLowerCase() === l
   ) || null;
 };
-
-// ─── Shared: hero ─────────────────────────────────────────────────────────────
-
-export function buildHero(user: any, isEditable: boolean, controls: any[] = []) {
-  if (!user) {
-    return (
-      <div className="AvocadoUserPage-hero AvocadoUserPage-hero--skeleton">
-        <div className="AvocadoUserPage-hero-inner">
-          <div className="AvocadoUserPage-hero-row">
-            <div className="AvocadoUserPage-shimmer AvocadoUserPage-shimmer--avatar" />
-            <div style={{ flex: 1 }}>
-              <div className="AvocadoUserPage-shimmer AvocadoUserPage-shimmer--name" />
-              <div className="AvocadoUserPage-shimmer AvocadoUserPage-shimmer--meta" />
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  const color           = user.color?.() || '#5a6480';
-  const badges          = user.badges?.().toArray?.() || [];
-  const isOnline        = user.isOnline?.();
-  const joinTime        = user.joinTime?.();
-  const joinLabel       = joinTime
-    ? new Date(joinTime as string).toLocaleDateString('en-US', { year: 'numeric', month: 'short', day: 'numeric' })
-    : null;
-
-  return (
-    <div className="AvocadoUserPage-hero" style={{ '--user-color': color }}>
-      <div className="AvocadoUserPage-hero-inner">
-        <div className="AvocadoUserPage-hero-row">
-          <div className="AvocadoUserPage-hero-avatarWrap">
-            {isEditable ? <AvatarEditor user={user} /> : <Avatar user={user} loading="eager" />}
-            {isOnline && <span className="AvocadoUserPage-onlineDot" />}
-          </div>
-          <div className="AvocadoUserPage-hero-info">
-            <h1 className="AvocadoUserPage-hero-name">{user.displayName?.() || user.username?.()}</h1>
-            {badges.length > 0 && (
-              <ul className="AvocadoUserPage-hero-badges badges">{listItems(badges)}</ul>
-            )}
-            <div className="AvocadoUserPage-hero-stats">
-              {isOnline && (
-                <span className="AvocadoUserPage-hero-statPill AvocadoUserPage-hero-statPill--online">
-                  <i className="fas fa-circle" aria-hidden="true" />Online
-                </span>
-              )}
-              {joinLabel && (
-                <span className="AvocadoUserPage-hero-statPill">Joined {joinLabel}</span>
-              )}
-            </div>
-          </div>
-          {controls.length > 0 && (
-            <div className="AvocadoUserPage-hero-controls">
-              <Dropdown
-                buttonClassName="Button AvocadoUserPage-controlsBtn"
-                menuClassName="Dropdown-menu--right"
-                label={app.translator.trans('core.forum.user_controls.button') as string}
-              >
-                {controls}
-              </Dropdown>
-            </div>
-          )}
-        </div>
-        {(() => {
-          try {
-            const UserBio = (flarum as any).reg.get('fof-user-bio', 'forum/components/UserBio');
-            if (UserBio && user.attribute('canViewBio')) {
-              return <div className="AvocadoUserPage-hero-bio"><UserBio user={user} editable={isEditable} /></div>;
-            }
-          } catch { }
-          return null;
-        })()}
-      </div>
-    </div>
-  );
-}
-
-// ─── Scrollable nav ───────────────────────────────────────────────────────────
-
-class ScrollableNav {
-  private _el: Element | null = null;
-  private _canLeft  = false;
-  private _canRight = false;
-  private _dragging = false;
-  private _startX   = 0;
-  private _scrollLeft0 = 0;
-  private _ro: ResizeObserver | null = null;
-  private _handleScroll: (() => void) | null = null;
-  private _handleMouseDown: ((e: MouseEvent) => void) | null = null;
-  private _handleMouseMove: ((e: MouseEvent) => void) | null = null;
-  private _handleMouseUp: (() => void) | null = null;
-
-  private _check() {
-    const el = this._el as HTMLElement;
-    if (!el) return;
-    const l = el.scrollLeft > 1;
-    const r = el.scrollLeft + el.clientWidth < el.scrollWidth - 1;
-    if (l !== this._canLeft || r !== this._canRight) {
-      this._canLeft  = l;
-      this._canRight = r;
-      m.redraw();
-    }
-  }
-
-  private _scroll(dir: number) {
-    (this._el as HTMLElement)?.scrollBy({ left: dir * 160, behavior: 'smooth' });
-  }
-
-  oncreate(vnode: any) {
-    const el = vnode.dom.querySelector('.AvocadoUserPage-navInner') as HTMLElement;
-    this._el = el;
-    if (!el) return;
-
-    this._handleScroll    = () => this._check();
-    this._handleMouseDown = (e: MouseEvent) => {
-      this._dragging    = true;
-      this._startX      = e.pageX - el.offsetLeft;
-      this._scrollLeft0 = el.scrollLeft;
-      document.documentElement.style.cursor     = 'grabbing';
-      document.documentElement.style.userSelect = 'none';
-    };
-    this._handleMouseMove = (e: MouseEvent) => {
-      if (!this._dragging) return;
-      e.preventDefault();
-      el.scrollLeft = this._scrollLeft0 - (e.pageX - el.offsetLeft - this._startX);
-    };
-    this._handleMouseUp = () => {
-      if (!this._dragging) return;
-      this._dragging = false;
-      document.documentElement.style.cursor     = '';
-      document.documentElement.style.userSelect = '';
-    };
-
-    el.addEventListener('scroll',     this._handleScroll!,    { passive: true });
-    el.addEventListener('mousedown',  this._handleMouseDown!);
-    window.addEventListener('mousemove', this._handleMouseMove!);
-    window.addEventListener('mouseup',   this._handleMouseUp!);
-
-    this._ro = new ResizeObserver(() => this._check());
-    this._ro.observe(el);
-    this._check();
-
-    const activeItem = el.querySelector('li.active') as HTMLElement | null;
-    if (activeItem) {
-      requestAnimationFrame(() => {
-        el.scrollLeft = activeItem.offsetLeft - (el.clientWidth - activeItem.clientWidth) / 2;
-      });
-    }
-  }
-
-  onremove() {
-    if (this._el) {
-      this._el.removeEventListener('scroll',    this._handleScroll!);
-      this._el.removeEventListener('mousedown', this._handleMouseDown!);
-    }
-    window.removeEventListener('mousemove', this._handleMouseMove!);
-    window.removeEventListener('mouseup',   this._handleMouseUp!);
-    this._ro?.disconnect();
-    this._el = null;
-  }
-
-  view(vnode: any) {
-    return (
-      <div className="AvocadoUserPage-nav">
-        <button
-          className={`AvocadoUserPage-navArrow AvocadoUserPage-navArrow--left${this._canLeft ? ' is-visible' : ''}`}
-          onclick={() => this._scroll(-1)}
-          aria-label="Scroll left"
-          tabindex="-1"
-        >
-          <i className="fas fa-chevron-left" aria-hidden="true" />
-        </button>
-        {vnode.children}
-        <button
-          className={`AvocadoUserPage-navArrow AvocadoUserPage-navArrow--right${this._canRight ? ' is-visible' : ''}`}
-          onclick={() => this._scroll(1)}
-          aria-label="Scroll right"
-          tabindex="-1"
-        >
-          <i className="fas fa-chevron-right" aria-hidden="true" />
-        </button>
-      </div>
-    );
-  }
-}
-
-// ─── Shared nav builders ──────────────────────────────────────────────────────
-
-export function buildSidebar(page: any) {
-  const user = page?.user;
-  if (!user) {
-    return (
-      <div className="AvocadoUserPage-nav">
-        <div className="AvocadoUserPage-navInner">
-          {[1, 2, 3, 4].map((i) => (
-            <div key={i} className="AvocadoUserPage-shimmer AvocadoUserPage-shimmer--navItem" />
-          ))}
-        </div>
-      </div>
-    );
-  }
-  return (
-    <ScrollableNav>
-      <ul className="AvocadoUserPage-navInner">
-        {listItems(page.navItems().toArray())}
-      </ul>
-    </ScrollableNav>
-  );
-}
-
-export function buildUserPhoneNav(page: any) {
-  const user  = page?.user;
-  const items = user ? page.navItems().toArray() : [];
-  return (
-    <nav className="IndexPage-nav sideNav">
-      <ul>
-        <li className="item item-nav">
-          <SelectDropdown className="App-titleControl" buttonClassName="Button">
-            {items}
-          </SelectDropdown>
-        </li>
-      </ul>
-    </nav>
-  );
-}
 
 // ─── Base page ────────────────────────────────────────────────────────────────
